[Xamarin.Android.Build.Tests] Fix Repeat Build Tests (#2582)
PR #2515 highlighted an issue with a few of our repeat build tests.
Tests were failing because the `_Sign` target was being skipped.
Looking at the build logs, the `BuildApk` target was being called,
as was the `_CreateBaseApk` target, so what is going on?

It turns out that in the `<BuildApk/>` task we use the
`MonoAndroidHelper.CopyIfZipChanged()` method to move the temp zip
file over to the actual one.  In this case, the zip was identical!
Thus the "new" zip was not copied, and the `_Sign` target did not
need to run.

How can the zip be the same?  Didn't it have an updated file?
Not quite; we just touched the timestamp.  In this case the zip ends
up with the exact same CRC values for each file.  We use the CRC to
detect changes, so no changes were detected.  It turns out our build
system was behaving as expected, it was just our test was invalid.

The fix here is to update the tests to make sure we do change a file.
This means the CRC's will change and the targets will run.