[Xamarin.Android.Build.Tasks] remove more %temp% file usage
In various places throughout our build, we are still using a pattern
such as:

    try {
        var temp = Path.GetTempFileName ();
        //Write to the file
        MonoAndroidHelper.CopyIfChanged (temp, dest);
    } finally {
        File.Delete (temp);
    }

Or sometimes variations that aren't even as nice as this...

We can instead use the new `MonoAndroidHelper.CopyIfStreamChanged`
method and not use any temp files.

~~ CopyResource ~~

This MSBuild task had its own `CopyIfStreamChanged` behavior, and it
has been around a while. I was able to remove a lot of code from this
MSBuild task; it is much simpler now.

Other changes:

- We don't really need an extra `Run` method here.
- If we are putting `Assembly.GetExecutingAssembly ()` in a static
  variable, let's put the jcwgen assembly in one as well.
- Use string interpolation: `$`

~~ GeneratePackageManagerJava, MonoAndroidHelper, and XDocumentExtensions ~~

Each had a simple example of using a temp file that could be changed
to use `CopyIfStreamChanged` instead.

~~ Results ~~

I couldn't really see a build performance improvement, but it should
be slightly better. Most of these changes result in cleaner code, and
less temp files!