[C++] Cosmetic changes to the code: NULL -> nullptr
In effort to morph our current code to modern C++, this commit replaces all the
`NULL` macro instances with the C++ standard `nullptr` type. No functional
changes are present in this commit.

The only exception here is the `xamarin_getifaddrs.cc` file in which the reverse
operation is done - replacement of `nullptr` with `NULL`. The reason for this is
that the file is destined to be moved to the Mono codebase at some point and so,
at least for now, it should remain standard C99