[Xamarin.Android.Build.Tasks]  Fix problem with `_CompileJava` running always.
Commit b90d3ab7 altered where we generate the Guid for the
`XAMARIN_BUILD_ID`. This change resulted in the `_CompileJava`
target running on every build. This was because the Guid and
the resultng `AndroidEnvironment.java` file changed on every
build.

The `XAMARIN_BUILD_ID` should only change when an apk is
built. This is so the same id can be written to both the
apk and the msym archive.

There was also a problem on windows where the design time
build would trigger `_GeneratePackageManagerJava` to run.
This was down to the the following file being updated

	<Project>.csproj.CoreCompileInputs.cache

This would cause a cascade because once `_GeneratePackageManagerJava`
ran, this would then trigger `_CompileJava` and then other
targets.

Looking at `GeneratePackageManagerJava` we don't actually
use the assemblies directly within the tast. We just need
a list of the assembly names. So allot of the time we are
running this task if an assembly changes when we don't
really need to. What we need to do is run the task when
assemblies and either added or removed from the list.

To that end we now store a Hash of the `@(ResolvedAssemblies)`
ItemGroup in `$(IntermediateOutputPath)resolvedassemblies.hash`.
This hash is updated when the `ResolveAssemblies` target is run.
We can then use that hash file as an input to
`_GeneratePackageManagerJava` to ensure that we only run that
target when the list changes.

An addition change was to write the `XAMARIN_BUILD_ID` to a file
so that it can be stored between builds. This is needed because
if the `_GeneratePackageManagerJava` is skipped (due to an up to
date build) during the package creation the property
`$(_XamarinBuildId)` would be empty. This is required for
`CreateMsymManifest`, so we need to be able to store this value
between builds and read it in when required. Reading
directly from `txt` file is quicker than pulling the information
from the `XamarinAndroidEnvironmentVariables.java` file
which `_GeneratePackageManagerJava` produces.