[C++ification] Making AndroidSystem class more C++ and faster
It's a start of effort to slowly (and piece by piece) modify our runtime code
away from the C syntax toward the C++ one, using idioms and constructs from the
latter wherever possible, practical or simply better. The goal is to have code
that's safer, more correct, faster as well as to expose only the minimum
"public" API of every class in the interest of minimizing the potential of code
abuse.

At the same time, the changes should not affect the way the code works on the
runtime (unless the change implements a performance increase)

Each commit/PR in this series will attempt to mimimize collateral changes to
other classes/parts of code, but sometimes (as in this case) there will be
minimal changes in other areas related to the main class changed by a commit.

The following changes to the `AndroidSystem` class are made in this commit:

  * The `BundledProperty` structure moves to `android-system.cc` - it has no use
    outside the class
  * Several `printf` family calls are replaced with either simple `strdup` or by
    compile-time string concatenation.
  * `for` loops use `size_t` type for the counter variable
  * `nullptr` is used in place of the `!pointer` or `pointer` in conditional
    expressions or asserts.
  * The `_monodroid__system_property_get` overload which uses `popen(3)` to
    retrieve the property is removed - the `__system_property_get` call is still
    present in the Android NDK (albeit deprecated) and since we're compiling
    against API16 at least, we're in no immediate danger of missing that call
  * Local function variables are moved closer to places where they are first
    initialized/used.
  * Replaced the `TRY_LIBMONOSGEN` macro with an inline function
  * Replaced use of `mono_bool` with the C++ type `bool`
  * Replaced a number of `malloc/free` calls with `new/delete`
  * The `MONO_SGEN_ARCH_SO` constant is now generated on the compile time
    instead of on the runtime.

Other changes:

  * Define the `ANDROID64` macro on build time if targetting 64-bit Android.
    This macro was mistakenly omitted in conversion to `cmake`
  * Fix a handful of warnings regarding our `new/delete` operator
    implementations.