Add support for deferred builds when the IDE supports them
In order to take the most expensive target in a DTB (`UpdateAndroidResources`) out
of the DTB, the IDE can monitor design-time builds and schedule a deferred build
to run after the DTB finishes.

When the IDE supports deferred builds, the  `DeferredBuildSupported` will be
true, and in that case, the `DeferredBuild` target will run after the DTB finishes. To
schedule targets to run as part of that deferred build, they need to be added to
the `DeferredBuildDependsOn` property.

Targets can selectively disable themselves on DTBs when the deferred build is
supported by checking for both `'$(DeferredBuildSupported)' == 'true'` and the
`'$(DeferredBuild)' == 'true'` property, which is set for those builds.

The deferred build queuing will be based on the following condition: a regular design
time build was triggered and the targets run were either "Compile" or any target
that contains "DesignTime" in its name (this excludes queuing a deferred build
when a target like `GetFrameworkPaths` or `GetTargetPath` is run in a DTB, for
example).

Note that `Inputs`/`Outputs` on the targets, as well as any `DependsOnTargets`
specified remain the same as before. The effect is that expensive targets that caused
a slow DTB before, can now be scheduled to run *after* the build finishes (this is
when the deferred build is queued).

It's entirely opt-in (if no targets are specified for `DeferredBuildDependsOn`
it will be essentially a no-op run).

Finally, when a regular build starts, the existing deferred build will be waited for
completion before continuing, to ensure no concurrency happens.

The initial `Compile` target is run by the DTB in VS while
skipping the compiler call, at which point the UpdateAndroidResources
should also be skipped when the `DeferredBuildSupported` is `True`.

Subsequently, we run the `DeferredBuild` (or rather, the `UpdateAndroidResources`
which would be run by the `DeferredBuildDependsOn` we append it to)
which should also succeed (it's run with the same global parameters/
properties as the initial DTB).

Finally, we run the normal build, at which point everything should
succeed like before.

If an XVS version that supports this feature is installed, you can turn it on
by setting an environment variable `DeferredBuildSupported=true`. An A/B experiment
or remote setting can also set this value.