[Xamarin.Android.Build.Tasks] Add `directBootAware` if needed.
Fixes #2081

If a user needs to use the `directBootAware` attribute on
an `application`, `activity` or `service` we also need
to include it on the `MonoRuntimeProvider` so that
native libraries can be resolved correctly. Not doing
so results in the following

	java.lang.UnsatisfiedLinkError: No implementation found for void mono.android.Runtime.register

This commit looks to see if any elements in the manifest
contain `directBootAware`. If there are , then we also
include it in the `provider`.