Add an Azure DevOps project to build, test, and package (#56)
An [Azure DevOps build pipeline][0] has been configured in the public
Xamarin project to build and test Xamarin.Android.Tools on both
Windows and macOS.  The steps and environments used in this new build
pipeline are defined in the new `azure-pipelines.yaml` file, and are
described below.  A build status badge has also been added to the
README.md file which links to the latest successful build.  This link
can be used to quickly fetch the latest NuGet package produced by
this build.

**Windows Steps**:

 1. Build `Xamarin.Android.Tools.sln`.
 2. Run all unit tests matching the pattern `*-Tests.dll`.
 3. Create a NuGet package based on Major, Minor, and Patch versions
    defined in `azure-pipelines.yaml`.  The Patch value is the
    number of commits since the `nuget.version` file changed.
 4. Saves the NuGet package as a build artifact.  This build
    artifact can be downloaded directly from the Azure DevOps build
    info page.  Additionally, these artifact can be published to a
    public NuGet feed by running the corresponding
    [Azure DevOps release pipeline][1].

**macOS Steps**

 1. Build `Xamarin.Android.Tools.sln`.
 2. Run all unit tests contained in
    `Xamarin.Android.Tools.AndroidSdk-Tests.dll`.

This PR also fixes a small bug in an existing unit test which was
uncovered by the environments used for the Azure DevOps macOS build
pipeline.  `JdkInfo.GetKnownSystemJdkInfos()` prefers `$JAVA_HOME`
over any valid JDK values in the `$PATH` environment variable.
In order for this test to accurately test initializing from `$PATH`,
`$JAVA_HOME` must not point to a valid JDK.

[0]: https://dev.azure.com/xamarin/Xamarin/_build?definitionId=2&_a=summary
[1]: https://dev.azure.com/xamarin/Xamarin/_release?definitionId=1&view=mine