Typename mapping + environment assembly generator
Xamarin.Android needs a lot of information on the runtime in order to do its
tasks. Among this is are two bits of information that are used in the runtime
startup: type mappings (java-to-managed and managed-to-java) and environment
settings (variables and system properties). So far, Xamarin.Android used these
bits of data (generated on the application build time) by loading them and
parsing dynamically during startup. However, none of this data is *actually*
dynamic - it's known beforehand and should be stored as static data right in the
application. The problem with this idea, however, was that we had no way to put
that information into the application native bits on the build time: it would
require the presence of a compiler (or native assembler) as well as the native
linker for the target platforms. Until recently, it meant having to have the
Android NDK installed, a pretty hefty requirement. However, right now Android
SDK ships the native linker with its `build-tools` package and we can bundle in
our packages the, relatively small, `gas` native assembler for all the
architectures we support which allows us to generate simple native assembler
code, compile it and link into a shared library **when packaging** the
application.

The way it works is that during Xamarin.Android build we also build a stub
`libxamarin-app.so` library which the XA runtime is linked against. That stub
library is **not** shipped with XA, instead the version we generate on the build
time is packaged together with the runtime inside the APK. Since it's
abi-compatible with the one built previously, the dynamic loader doesn't care
and loads everything just fine.

This allows us to dispense with memory allocation, loading (potentially large)
files from the apk (lots of I/O, slow), parsing text files to set environment
variables (this also includes memory allocation). All of the data is loaded for
us by Android and we simply access global variables from the `libxamarin-app.so`
library.

The process, however, can be time consuming (sort of...) and in order to allow
fast deployment to work as quickly as possible, the debug (and desktop/host -
for the benefit of the UI Designer) builds of the runtime also contain code to
load both type maps and environment variables from files stored inside one of
the override directories.

The changes got us a speed-up of ~20ms for a release build of a Xamarin.Forms
app on Pixel 3 XL.