Fix build on dev16 by targeting net472
[Visual Studio 2019][0] has internally moved to target .NET 4.7.2,
and thus all the Visual Studio-provided MSBuild assemblies such as
`Microsoft.Build.dll` now target .NET 4.7.2, which in turn requires
that `$(TargetFrameworkVersion)`=v4.7.2 in order to reference the
Visual Studio-provided MSBuild assemblies.  If an earlier
`$(TargetFrameworkVersion)` value is used, then an MSB3274 warning
is generated:

	warning MSB3274: The primary reference "Microsoft.Build" could not be resolved because it was built against the ".NETFramework,Version=v4.7.2" framework.
	This is a higher version than the currently targeted framework ".NETFramework,Version=v4.6.2".

This subsequently turns into a CS0234 error:

	error CS0234: The type or namespace name 'Build' does not exist in the namespace 'Microsoft' (are you missing an assembly reference?)

Support building the xamarin-android repo by bumping the
`$(TargetFrameworkVersion)` value of projects which consume the
MSBuild assemblies to `v4.7.2`, avoiding the above warnings/errors,
but only when `$(VisualStudioVersion)` is >= 16.0, so that we can
still build the xamarin-android repo with current stable versions of
Visual Studio 2017.

[0]: https://web.archive.org/web/20190112190709/https://visualstudio.microsoft.com/vs/preview/