[Xamarin.Android.Build.Tasks] Run ILRepack if `NuGet.ProjectModel` Exists
We saw the following error in the latest versions of monodroid.

	/Library/Frameworks/Mono.framework/External/xbuild/Xamarin/Android/Xamarin.Android.Common.targets(1928,2): error : Could not load file or assembly 'NuGet.ProjectModel, Version=4.6.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35' or one of its dependencies.

After investingating it turns out that the
`Xamarin.Android.Build.Tasks` project was being run
multiple times (as expected). However the
`CopyFilesToOutputDirectory` target was being run
on the second build. This was also running the
internal target `_CopyFilesMarkedCopyLocal`.

The result of both of these targets was to copy
over the old assembly over the ILRepacked version
and copy over ALL the NuGet assemblies from the
nuget packages cache.

This then turns into the above error at build time
because the system has a different version of Nuget.
This was why we were using ILReapck in the first place.

We tried a number of different solutions. One was to update
the file in the intermediate directory which looked like
it would solve the problem initially. However in that
instance the Nuget files were still copied over to the
`$(OutputPath)` and might cause problems in the final
release. There was also another problem with ILRepack.
If we fixed up the file in the `$(IntermediateOutputPath)`
before `CopyFilesToOutputDirectory` we can end up with
ILRepack running again on the same assembly. This causes it
to pull in ALL the Nuget stuff again and we end up with a
5meg file rather than a 3meg one... Which was weird.

So the simplest solutiuon is to run ILRepack after
the `CopyFilesToOutputDirectory` target and only if
the `Nuget.ProjectModel` assembly exists in the `$(OutputPath)`.
Since the `ILRepacker` target is the one that deletes
the Nuget assemblies we can be sure that if they exist
we need to fix up the main assembly.
While this might mean that ILRepack runs twice in a build,
it does mean we get consistent output.