[Xamarin.Android.Build.Tasks] <GenerateJavaStubs/> only for MonoAndroid assemblies
Currently the `<GenerateJavaStubs/>` MSBuild task will load and
iterate over NetStandard or PCL assemblies. It simply looks at all
items in `@(_ResolvedUserAssemblies)` and writes Java stubs if needed.

We could look at the `TargetFrameworkIdentifier` and *only* do this
work for `MonoAndroid` assemblies.

To make this work:

- The `<ResolveAssemblies/>` MSBuild task needs to parse the
  `TargetFrameworkIdentifier` and add it to the item metadata for each
  assembly.
- `<GenerateJavaStubs/>` should use a subset of
  `@(_ResolvedUserAssemblies)` where `TargetFrameworkIdentifier` is
  `MonoAndroid`.

Results:

    Before
        First:       1200 ms  GenerateJavaStubs                          1 calls
        Incremental: 1047 ms  GenerateJavaStubs                          1 calls
    After
        First:       1084 ms  GenerateJavaStubs                          1 calls
        Incremental:  868 ms  GenerateJavaStubs                          1 calls

This seems to save ~200ms in the Xamarin.Forms integration project in
this repo. Projects using lots of NetStandard/PCL assemblies should
see a bigger benefit.