[build] dependencies.csproj before NuGet restore
Context: https://github.com/xamarin/xamarin-android/issues/2562

Downstream in monodroid, we got a failure such as:

    msbuild /p:Configuration=Release /p:AutoProvision=True /p:AutoProvisionUsesSudo=True /p:IgnoreMaxMonoVersion=False /p:XARepositoryName=monodroid /p:XAVersionCommitCount=3 /p:ProductVersion=9.2.99 /p:XAVersionHash=48c518f4d /p:XAVersionBranch=HEAD /v:d /t:Restore Java.Interop.sln
    ...
    external/xamarin-android/external/Java.Interop/src/Java.Interop/Java.Interop.csproj : error MSB4244: The SDK resolver assembly "/Library/Frameworks/Mono.framework/Versions/5.16.0/lib/mono/msbuild/15.0/bin/SdkResolvers/Microsoft.Build.NuGetSdkResolver/..\..\Microsoft.Build.NuGetSdkResolver.dll" could not be loaded. Invalid Image

My guess here (similar to #2562), is that we need to upgrade Mono
before `msbuild Java.Interop.sln /t:Restore` will work.

Luckily, this command is currently working *before* doing a NuGet
restore:

    msbuild build-tools/dependencies/dependencies.csproj

This means we can just reorder the targets in the `Makefile` so we
build `dependencies.csproj` before doing all the NuGet restores. This
is a bit simpler than the rework that would be required to bootstrap
Mono outside of MSBuild as described in #2562.